Implement sources reading!
This gives panicparse superpowers. It is able to deduce the types and be much
smarter in the way it present the information.

The code is still brittle and expected to fail in exotic cases. Please report
failures upstream. This is why it is not enabled by default, one has to use the
-parse flag to use this functionality. Eventually the flag will be enabled by
default and a -no-parse will be added.